+++ src/main/java/org/jungppo/bambooforest/battery/presentation/BatteryController.java
@@ -1,11 +1,11 @@
-package org.jungppo.bambooforest.controller.battery;
+package org.jungppo.bambooforest.battery.presentation;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import org.jungppo.bambooforest.dto.battery.BatteryItemDto;
-import org.jungppo.bambooforest.entity.battery.BatteryItem;
+import org.jungppo.bambooforest.battery.domain.BatteryItem;
+import org.jungppo.bambooforest.battery.dto.BatteryItemDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
